+++ src/lektorium/repo/__init__.py
@@ -1,0 +1,2 @@
+from .memory import Repo as ListRepo, SITES
+from .git import Repo as GitRepo
