+++ src/lektorium/repo/interface.py
@@ -1,0 +1,8 @@
+import abc
+
+
+class Repo:
+    @property
+    @abc.abstractmethod
+    def sites(self):
+        pass
